requirements.txt
Flask
keras
numpy
pillow